requirements.txt
#
# This file is autogenerated by pip-compile with python 3.9
# To update, run:
#
#    pip-compile requirements.in
#
anyio==3.6.2
    # via starlette
click==8.1.3
    # via uvicorn
fastapi==0.85.1
    # via -r requirements.in
h11==0.14.0
    # via uvicorn
idna==3.4
    # via anyio
ml-base==0.2.1
    # via -r requirements.in
pydantic==1.10.2
    # via
    #   fastapi
    #   ml-base
pyyaml==6.0
    # via -r requirements.in
sniffio==1.3.0
    # via anyio
starlette==0.20.4
    # via fastapi
typing-extensions==4.4.0
    # via
    #   pydantic
    #   starlette
uvicorn==0.18.3
    # via -r requirements.in
